pyproject.toml
[tool.poetry]
name = "inflation-dashboard"
version = "1.0.0"
description = ""
authors = ["Abrar Mudhir <abrarmudhir@outlook.com>"]
readme = "README.md"
packages=[{ include="app"},]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
dash = "^2.17.1"
pandas = "^2.2.2"
a2wsgi = "^1.10.4"
dash-bootstrap-components = "^1.6.0"
pydantic-settings = "^2.3.3"
sqlalchemy = "^2.0.31"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
